+++ applications/shared/view/timeclock/settings.php
@@ -24,51 +24,27 @@
         <div class="well">
             <form class="form-horizontal" name="settings_form" method="post" action="">
                 <div class="group">
-                    <label class="control-label">Pay Period Start</label>
-                    <select class="form-control" name="pay_period_start">
-                        <option value="Sunday">Sunday</option>
-                        <option value="Monday" selected="selected">Monday</option>
-                        <option value="Tuesday">Tuesday</option>
-                        <option value="Wednesday">Wednesday</option>
-                        <option value="Thursday">Thursday</option>
-                        <option value="Friday">Friday</option>
-                        <option value="Saturday">Saturday</option>
-                    </select>
-                </div>
-                <div class="group">
-                    <label>Pay Period End</label>
-                    <select class="form-control">
-                        <option value="Sunday">Sunday</option>
-                        <option value="Monday">Monday</option>
-                        <option value="Tuesday">Tuesday</option>
-                        <option value="Wednesday">Wednesday</option>
-                        <option value="Thursday">Thursday</option>
-                        <option value="Friday">Friday</option>
-                        <option value="Saturday">Saturday</option>
-                    </select>
-                </div>
-                <div class="group">
                     <label>Round Time By</label>
-                    <select class="form-control">
-                        <option value="none">None</option>
-                        <option value="1">1 Minute</option>
-                        <option value="15">15 Minutes</option>
-                        <option value="30">30 Minutes</option>
+                    <select class="form-control" name="round_time_by">
+                        <option value="none" <?php echo ('none' === $this->system_di->template->round_time_by) ? 'selected="selected"' : ''; ?>>None</option>
+                        <option value="1" <?php echo ('1' === $this->system_di->template->round_time_by) ? 'selected="selected"' : ''; ?>>1 Minute</option>
+                        <option value="15" <?php echo ('15' === $this->system_di->template->round_time_by) ? 'selected="selected"' : ''; ?>>15 Minutes</option>
+                        <option value="30" <?php echo ('30' === $this->system_di->template->round_time_by) ? 'selected="selected"' : ''; ?>>30 Minutes</option>
                     </select>
                 </div>
                 <div class="group">
                     <label>Sort Employees By</label>
-                    <select class="form-control">
-                        <option value="last_first">Last Name, First Name</option>
-                        <option value="first_last">First Name, Last Name</option>
-                        <option value="uid">UID</option>
+                    <select class="form-control" name="sort_employees_by">
+                        <option value="last_name" <?php echo ('last_name' === $this->system_di->template->sort_employees_by) ? 'selected="selected"' : ''; ?>>Last Name, First Name</option>
+                        <option value="first_name" <?php echo ('first_name' === $this->system_di->template->sort_employees_by) ? 'selected="selected"' : ''; ?>>First Name, Last Name</option>
+                        <option value="uid" <?php echo ('uid' === $this->system_di->template->sort_employees_by) ? 'selected="selected"' : ''; ?>>UID</option>
                     </select>
                 </div>
                 <div class="group">
                     <label>List Employees As</label>
-                    <select class="form-control">
-                        <option value="last_first">Last Name, First Name</option>
-                        <option value="first_last">First Name, Last Name</option>
+                    <select class="form-control" name="list_employees_as">
+                        <option value="last_first" <?php echo ('last_first' === $this->system_di->template->list_employees_by) ? 'selected="selected"' : ''; ?>>Last Name, First Name</option>
+                        <option value="first_last"<?php echo ('first_last' === $this->system_di->template->list_employees_by) ? 'selected="selected"' : ''; ?>>First Name, Last Name</option>
                     </select>
                 </div>
                 {update_status}